Fixes. Working on packaging the entire thing together. Release might be slightly delayed.
Some things have internal buffers that I'm not in control of, so some parts of the translation had to be cut (DICT transcriptions, DATA couple of messages are slightly trimmed)

More typo fixes

Apparently topographer had a bug, where in some cases it would put a wrong pointer. Fixed it by saying that the location of the pointer cannot be odd.